+++ htmldoc/ps-pdf.cxx
@@ -580,17 +580,14 @@ pspdf_export_out(tree_t *document,	/* I - Document to export */
       // Find the title file...
       if ((title_file = file_find(Path, TitleImage)) == NULL)
       {
-	progress_error(HD_ERROR_FILE_NOT_FOUND,
-	               "Unable to find title file \"%s\"!", TitleImage);
+	progress_error(HD_ERROR_FILE_NOT_FOUND, "Unable to find title file '%s'.", TitleImage);
 	return (1);
       }
 
       // Write a title page from HTML source...
       if ((fp = fopen(title_file, "rb")) == NULL)
       {
-	progress_error(HD_ERROR_FILE_NOT_FOUND,
-	               "Unable to open title file \"%s\" - %s!",
-                       TitleImage, strerror(errno));
+	progress_error(HD_ERROR_FILE_NOT_FOUND, "Unable to open title file '%s': %s", TitleImage, strerror(errno));
 	return (1);
       }
 
@@ -1222,12 +1219,17 @@ pspdf_prepare_outpages()
 
 
   // Allocate an output page array...
-  outpages = (outpage_t *)malloc(sizeof(outpage_t) * num_pages);
+  num_outpages = 0;
+
+  if ((outpages = (outpage_t *)malloc(sizeof(outpage_t) * num_pages)) == NULL)
+  {
+    progress_error(HD_ERROR_OUT_OF_MEMORY, "Unable to allocate memory for %d output pages.", (int)num_pages);
+    return;
+  }
 
   memset(outpages, -1, sizeof(outpage_t) * num_pages);
 
-  num_outpages = 0;
-  outpage      = outpages;
+  outpage = outpages;
 
   // Handle the title page, as needed...
   if (TitlePage)
@@ -2921,26 +2923,20 @@ pdf_write_contents(FILE   *out,			/* I - Output file */
 
   if ((entry_counts = (int *)calloc(sizeof(int), num_headings + 1)) == NULL)
   {
-    progress_error(HD_ERROR_OUT_OF_MEMORY,
-                   "Unable to allocate memory for %d headings - %s",
-                   (int)num_headings, strerror(errno));
+    progress_error(HD_ERROR_OUT_OF_MEMORY, "Unable to allocate memory for %d headings.", (int)num_headings);
     return;
   }
 
   if ((entry_objects = (int *)calloc(sizeof(int), num_headings + 1)) == NULL)
   {
-    progress_error(HD_ERROR_OUT_OF_MEMORY,
-                   "Unable to allocate memory for %d headings - %s",
-                   (int)num_headings, strerror(errno));
+    progress_error(HD_ERROR_OUT_OF_MEMORY, "Unable to allocate memory for %d headings.", (int)num_headings);
     free(entry_counts);
     return;
   }
 
   if ((entries = (tree_t **)calloc(sizeof(tree_t *), num_headings + 1)) == NULL)
   {
-    progress_error(HD_ERROR_OUT_OF_MEMORY,
-                   "Unable to allocate memory for %d headings - %s",
-                   (int)num_headings, strerror(errno));
+    progress_error(HD_ERROR_OUT_OF_MEMORY, "Unable to allocate memory for %d headings.", (int)num_headings);
     free(entry_objects);
     free(entry_counts);
     return;
@@ -3205,9 +3201,7 @@ pdf_start_object(FILE *out,	// I - File to write to
 
     if (temp == NULL)
     {
-      progress_error(HD_ERROR_OUT_OF_MEMORY,
-                     "Unable to allocate memory for %d objects - %s",
-                     (int)alloc_objects, strerror(errno));
+      progress_error(HD_ERROR_OUT_OF_MEMORY, "Unable to allocate memory for %d objects.", (int)alloc_objects);
       alloc_objects -= ALLOC_OBJECTS;
       return (0);
     }
@@ -3402,9 +3396,7 @@ pdf_write_links(FILE *out)		/* I - Output file */
 
   if ((lobjs = (int *)malloc(sizeof(int) * (size_t)alloc_lobjs)) == NULL)
   {
-    progress_error(HD_ERROR_OUT_OF_MEMORY,
-                   "Unable to allocate memory for %d link objects - %s",
-                   alloc_lobjs, strerror(errno));
+    progress_error(HD_ERROR_OUT_OF_MEMORY, "Unable to allocate memory for %d link objects.", alloc_lobjs);
     return;
   }
 
@@ -3526,9 +3518,7 @@ pdf_write_links(FILE *out)		/* I - Output file */
 		write_string(out, r->data.link, 0);
 
 		if (StrictHTML)
-		  progress_error(HD_ERROR_UNRESOLVED_LINK,
-		                 "Unable to resolve link to \"%s\"!",
-		                 r->data.link);
+		  progress_error(HD_ERROR_UNRESOLVED_LINK, "Unable to resolve link to '%s'.", r->data.link);
               }
 	    }
 	    else
@@ -4121,9 +4111,7 @@ parse_doc(tree_t *t,		/* I - Tree to parse */
       chapter ++;
       if (chapter >= MAX_CHAPTERS)
       {
-	progress_error(HD_ERROR_TOO_MANY_CHAPTERS,
-	               "Too many chapters/files in document (%d > %d)!",
-	               chapter, MAX_CHAPTERS);
+	progress_error(HD_ERROR_TOO_MANY_CHAPTERS, "Too many chapters/files in document (%d > %d).", chapter, MAX_CHAPTERS);
         chapter = MAX_CHAPTERS - 1;
       }
       else
@@ -4737,9 +4725,7 @@ parse_heading(tree_t *t,	/* I - Tree to parse */
 
       if (temp == NULL)
       {
-        progress_error(HD_ERROR_OUT_OF_MEMORY,
-                       "Unable to allocate memory for %d headings - %s",
-	               (int)alloc_headings, strerror(errno));
+        progress_error(HD_ERROR_OUT_OF_MEMORY, "Unable to allocate memory for %d headings.", (int)alloc_headings);
 	alloc_headings -= ALLOC_HEADINGS;
 	return;
       }
@@ -4756,9 +4742,7 @@ parse_heading(tree_t *t,	/* I - Tree to parse */
 
       if (temp == NULL)
       {
-        progress_error(HD_ERROR_OUT_OF_MEMORY,
-                       "Unable to allocate memory for %d headings - %s",
-	               (int)alloc_headings, strerror(errno));
+        progress_error(HD_ERROR_OUT_OF_MEMORY, "Unable to allocate memory for %d headings.", (int)alloc_headings);
 	alloc_headings -= ALLOC_HEADINGS;
 	return;
       }
@@ -5364,9 +5348,7 @@ parse_paragraph(tree_t *t,	/* I - Tree to parse */
 
 	    if (((temp->width - right + left) > 0.001 ||
 	         (temp->height - top + bottom) > 0.001)  && OverflowErrors)
-	      progress_error(HD_ERROR_CONTENT_TOO_LARGE,
-	                     "Text on page %d too large - "
-			     "truncation or overlapping may occur!", *page + 1);
+	      progress_error(HD_ERROR_CONTENT_TOO_LARGE, "Text on page %d too large -  truncation or overlapping may occur.", *page + 1);
 
             if (linetype == NULL)
             {
@@ -5407,9 +5389,7 @@ parse_paragraph(tree_t *t,	/* I - Tree to parse */
 	                    temp->width, temp->height,
 			    right - left, top - bottom));
 
-	      progress_error(HD_ERROR_CONTENT_TOO_LARGE,
-	                     "Image on page %d too large - "
-			     "truncation or overlapping may occur!", *page + 1);
+	      progress_error(HD_ERROR_CONTENT_TOO_LARGE, "Image on page %d too large - truncation or overlapping may occur.", *page + 1);
             }
 
 	    if ((border = htmlGetVariable(temp, (uchar *)"BORDER")) != NULL)
@@ -5752,9 +5732,7 @@ parse_pre(tree_t *t,		/* I - Tree to parse */
     }
 
     if ((*x - right) > 0.001 && OverflowErrors)
-      progress_error(HD_ERROR_CONTENT_TOO_LARGE,
-	             "Preformatted text on page %d too long - "
-		     "truncation or overlapping may occur!", *page + 1);
+      progress_error(HD_ERROR_CONTENT_TOO_LARGE, "Preformatted text on page %d too long -  truncation or overlapping may occur.", *page + 1);
 
     *y -= _htmlSpacings[t->size] - _htmlSizes[t->size];
   }
@@ -6650,18 +6628,16 @@ parse_table(tree_t *t,			// I - Tree to parse
 	else
 	  cells = (tree_t ***)realloc(cells, sizeof(tree_t **) * (size_t)alloc_rows);
 
-        if (cells == (tree_t ***)0)
+        if (cells == NULL)
 	{
-	  progress_error(HD_ERROR_OUT_OF_MEMORY,
-                         "Unable to allocate memory for table!");
+	  progress_error(HD_ERROR_OUT_OF_MEMORY, "Unable to allocate memory for table.");
 	  return;
 	}
       }
 
       if ((cells[table.num_rows] = (tree_t **)calloc(sizeof(tree_t *), MAX_COLUMNS)) == NULL)
       {
-	progress_error(HD_ERROR_OUT_OF_MEMORY,
-                       "Unable to allocate memory for table!");
+	progress_error(HD_ERROR_OUT_OF_MEMORY, "Unable to allocate memory for table.");
         free(cells);
 	return;
       }
@@ -7096,7 +7072,7 @@ parse_table(tree_t *t,			// I - Tree to parse
   }
 
   if ((width - right + left) > 0.001f && OverflowErrors)
-    progress_error(HD_ERROR_CONTENT_TOO_LARGE, "Table on page %d too wide - truncation or overlapping may occur!", *page + 1);
+    progress_error(HD_ERROR_CONTENT_TOO_LARGE, "Table on page %d too wide - truncation or overlapping may occur.", *page + 1);
 
   DEBUG_puts("");
 
@@ -8357,8 +8333,7 @@ parse_comment(tree_t *t,	/* I - Tree to parse */
       }
       else
       {
-        progress_error(HD_ERROR_BAD_COMMENT,
-                       "Bad HEADER position: \"%s\"", comment);
+        progress_error(HD_ERROR_BAD_COMMENT, "Bad HEADER position '%s'.", comment);
 	return;
       }
 
@@ -8367,8 +8342,7 @@ parse_comment(tree_t *t,	/* I - Tree to parse */
 
       if (*comment != '\"')
       {
-        progress_error(HD_ERROR_BAD_COMMENT,
-                       "Bad HEADER string: \"%s\"", comment);
+        progress_error(HD_ERROR_BAD_COMMENT, "Bad HEADER string '%s'.", comment);
 	return;
       }
 
@@ -8468,8 +8442,7 @@ parse_comment(tree_t *t,	/* I - Tree to parse */
       }
       else
       {
-        progress_error(HD_ERROR_BAD_COMMENT,
-                       "Bad HEADER1 position: \"%s\"", comment);
+        progress_error(HD_ERROR_BAD_COMMENT, "Bad HEADER1 position '%s'.", comment);
 	return;
       }
 
@@ -8478,8 +8451,7 @@ parse_comment(tree_t *t,	/* I - Tree to parse */
 
       if (*comment != '\"')
       {
-        progress_error(HD_ERROR_BAD_COMMENT,
-                       "Bad HEADER1 string: \"%s\"", comment);
+        progress_error(HD_ERROR_BAD_COMMENT, "Bad HEADER1 string: '%s'.", comment);
 	return;
       }
 
@@ -8572,8 +8544,7 @@ parse_comment(tree_t *t,	/* I - Tree to parse */
       }
       else
       {
-        progress_error(HD_ERROR_BAD_COMMENT,
-                       "Bad FOOTER position: \"%s\"", comment);
+        progress_error(HD_ERROR_BAD_COMMENT, "Bad FOOTER position: '%s'.", comment);
 	return;
       }
 
@@ -8582,8 +8553,7 @@ parse_comment(tree_t *t,	/* I - Tree to parse */
 
       if (*comment != '\"')
       {
-        progress_error(HD_ERROR_BAD_COMMENT,
-                       "Bad FOOTER string: \"%s\"", comment);
+        progress_error(HD_ERROR_BAD_COMMENT, "Bad FOOTER string: '%s'.", comment);
 	return;
       }
 
@@ -8865,7 +8835,9 @@ new_render(int      page,		/* I - Page number (0-n) */
   }
 
   if ((type != RENDER_TEXT && type != RENDER_LINK) || data == NULL)
+  {
     r = (render_t *)calloc(sizeof(render_t), 1);
+  }
   else
   {
     datalen = strlen((char *)data);
@@ -8874,8 +8846,7 @@ new_render(int      page,		/* I - Page number (0-n) */
 
   if (r == NULL)
   {
-    progress_error(HD_ERROR_OUT_OF_MEMORY,
-                   "Unable to allocate memory on page %d\n", (int)page + 1);
+    progress_error(HD_ERROR_OUT_OF_MEMORY, "Unable to allocate memory on page %d.", (int)page + 1);
     memset(&dummy, 0, sizeof(dummy));
     return (&dummy);
   }
@@ -8976,9 +8947,7 @@ check_pages(int page)	// I - Current page
 
     if (temp == NULL)
     {
-      progress_error(HD_ERROR_OUT_OF_MEMORY,
-                     "Unable to allocate memory for %d pages - %s",
-	             (int)alloc_pages, strerror(errno));
+      progress_error(HD_ERROR_OUT_OF_MEMORY, "Unable to allocate memory for %d pages.", (int)alloc_pages);
       alloc_pages -= ALLOC_PAGES;
       return;
     }
@@ -9080,7 +9049,7 @@ add_link(tree_t *html,		/* I - HTML node */
 
     if (temp == NULL)
     {
-      progress_error(HD_ERROR_OUT_OF_MEMORY, "Unable to allocate memory for %d links - %s", (int)alloc_links, strerror(errno));
+      progress_error(HD_ERROR_OUT_OF_MEMORY, "Unable to allocate memory for %d links.", (int)alloc_links);
       alloc_links -= ALLOC_LINKS;
       return;
     }
@@ -9722,7 +9691,12 @@ flatten_tree(tree_t *t)		/* I - Markup tree to flatten */
       case MARKUP_BR :
       case MARKUP_SPACER :
       case MARKUP_IMG :
-	  temp = (tree_t *)calloc(sizeof(tree_t), 1);
+	  if ((temp = (tree_t *)calloc(sizeof(tree_t), 1)) == NULL)
+	  {
+	    progress_error(HD_ERROR_OUT_OF_MEMORY, "Unable to allocate memory for node.");
+	    break;
+	  }
+
 	  memcpy(temp, t, sizeof(tree_t));
 	  temp->parent = parent;
 	  temp->child  = NULL;
@@ -9739,7 +9713,12 @@ flatten_tree(tree_t *t)		/* I - Markup tree to flatten */
       case MARKUP_A :
           if (htmlGetVariable(t, (uchar *)"NAME") != NULL)
           {
-	    temp = (tree_t *)calloc(sizeof(tree_t), 1);
+	    if ((temp = (tree_t *)calloc(sizeof(tree_t), 1)) == NULL)
+	    {
+	      progress_error(HD_ERROR_OUT_OF_MEMORY, "Unable to allocate memory for node.");
+	      break;
+	    }
+
 	    memcpy(temp, t, sizeof(tree_t));
 	    temp->parent = parent;
 	    temp->child  = NULL;
@@ -9778,7 +9757,12 @@ flatten_tree(tree_t *t)		/* I - Markup tree to flatten */
       case MARKUP_DT :
       case MARKUP_TR :
       case MARKUP_CAPTION :
-	  temp = (tree_t *)calloc(sizeof(tree_t), 1);
+	  if ((temp = (tree_t *)calloc(sizeof(tree_t), 1)) == NULL)
+	  {
+	    progress_error(HD_ERROR_OUT_OF_MEMORY, "Unable to allocate memory for node.");
+	    break;
+	  }
+
 	  temp->markup = MARKUP_BR;
 	  temp->parent = parent;
 	  temp->child  = NULL;
@@ -10550,6 +10534,12 @@ write_image(FILE     *out,		/* I - Output file */
     indices  = (uchar *)calloc((size_t)indwidth, (size_t)(img->height + 1));
 					// height + 1 for PS odd-row-count bug
 
+    if (indices == NULL)
+    {
+      progress_error(HD_ERROR_OUT_OF_MEMORY, "Unable to allocate memory for image '%s'.", img->filename);
+      return;
+    }
+
     if (img->depth == 1)
     {
      /*
@@ -11053,22 +11043,27 @@ write_image(FILE     *out,		/* I - Output file */
 
 	    if (img->mask && img->maskscale == 8)
 	    {
-	      data = (uchar *)malloc((size_t)(img->width * 2));
-
-	      for (i = 0, maskptr = img->mask, indptr = indices;
-	           i < img->height;
-		   i ++)
+	      if ((data = (uchar *)malloc((size_t)(img->width * 2))) == NULL)
 	      {
-	        for (j = img->width, dataptr = data; j > 0; j --)
+	        progress_error(HD_ERROR_OUT_OF_MEMORY, "Unable to allocate memory for '%s'.", img->filename);
+	      }
+	      else
+	      {
+		for (i = 0, maskptr = img->mask, indptr = indices;
+		     i < img->height;
+		     i ++)
 		{
-		  *dataptr++ = *maskptr++;
-		  *dataptr++ = *indptr++;
+		  for (j = img->width, dataptr = data; j > 0; j --)
+		  {
+		    *dataptr++ = *maskptr++;
+		    *dataptr++ = *indptr++;
+		  }
+
+		  flate_write(out, data, img->width * 2);
 		}
 
-		flate_write(out, data, img->width * 2);
+		free(data);
 	      }
-
-	      free(data);
 	    }
 	    else
 	      flate_write(out, indices, indwidth * img->height);
@@ -11133,35 +11128,40 @@ write_image(FILE     *out,		/* I - Output file */
 
 	    if (img->mask && img->maskscale == 8)
 	    {
-	      data = (uchar *)malloc((size_t)(img->width * (img->depth + 1)));
-
-	      for (i = 0, maskptr = img->mask, pixel = img->pixels;
-	           i < img->height;
-		   i ++)
+	      if ((data = (uchar *)malloc((size_t)(img->width * (img->depth + 1)))) == NULL)
 	      {
-	        if (img->depth == 1)
-		{
-	          for (j = img->width, dataptr = data; j > 0; j --)
-		  {
-		    *dataptr++ = *maskptr++;
-		    *dataptr++ = *pixel++;
-		  }
-		}
-		else
-		{
-	          for (j = img->width, dataptr = data; j > 0; j --)
-		  {
-		    *dataptr++ = *maskptr++;
-		    *dataptr++ = *pixel++;
-		    *dataptr++ = *pixel++;
-		    *dataptr++ = *pixel++;
-		  }
-		}
-
-		flate_write(out, data, img->width * (img->depth + 1));
+		progress_error(HD_ERROR_OUT_OF_MEMORY, "Unable to allocate memory for '%s'.", img->filename);
 	      }
+	      else
+	      {
+		for (i = 0, maskptr = img->mask, pixel = img->pixels;
+		     i < img->height;
+		     i ++)
+		{
+		  if (img->depth == 1)
+		  {
+		    for (j = img->width, dataptr = data; j > 0; j --)
+		    {
+		      *dataptr++ = *maskptr++;
+		      *dataptr++ = *pixel++;
+		    }
+		  }
+		  else
+		  {
+		    for (j = img->width, dataptr = data; j > 0; j --)
+		    {
+		      *dataptr++ = *maskptr++;
+		      *dataptr++ = *pixel++;
+		      *dataptr++ = *pixel++;
+		      *dataptr++ = *pixel++;
+		    }
+		  }
 
-	      free(data);
+		  flate_write(out, data, img->width * (img->depth + 1));
+		}
+
+		free(data);
+	      }
 	    }
 	    else
 	      flate_write(out, img->pixels,
@@ -11824,9 +11824,7 @@ write_prolog(FILE  *out,		/* I - Output file */
       }
       else
       {
-	progress_error(HD_ERROR_FILE_NOT_FOUND,
-                       "Unable to open data file \"%s\" - %s", temp,
-                       strerror(errno));
+	progress_error(HD_ERROR_FILE_NOT_FOUND, "Unable to open data file '%s': %s", temp, strerror(errno));
 
 	fprintf(out, "%%%%BeginResource: procset htmldoc-device 1.9 %s\n", version + 4);
 	fputs("languagelevel 1 eq{/setpagedevice{pop}BD}if\n", out);
@@ -12600,8 +12598,7 @@ write_type1(FILE       *out,		/* I - File to write to */
   if ((fp = fopen(filename, "r")) == NULL)
   {
 #ifndef DEBUG
-    progress_error(HD_ERROR_FILE_NOT_FOUND,
-                   "Unable to open font file %s!", filename);
+    progress_error(HD_ERROR_FILE_NOT_FOUND, "Unable to open font file '%s'.", filename);
 #endif /* !DEBUG */
     return (0);
   }
@@ -12722,8 +12719,7 @@ write_type1(FILE       *out,		/* I - File to write to */
     if ((fp = fopen(filename, "r")) == NULL)
     {
 #ifndef DEBUG
-      progress_error(HD_ERROR_FILE_NOT_FOUND,
-                     "Unable to open font width file %s!", filename);
+      progress_error(HD_ERROR_FILE_NOT_FOUND, "Unable to open font width file '%s'.", filename);
 #endif /* !DEBUG */
       return (0);
     }